+++ CHANGELOG.md
@@ -20,6 +20,7 @@ Compatibility:
 * Fixed `Socket#(local|remote)_address` to retrieve family and type from the file descriptor (#2444, @larskanis).
 * Add `Thread.ignore_deadlock` accessor (#2453).
 * Allow `Hash#transform_keys` to take a hash argument (@ccocchi, #2464).
+* Implement `--backtrace-limit` option (#2453).
 
 Performance: